refactor(sensors): make data storage more consistent
all data is now stored inside of the datapoint struct instead of in both
an unused "data" struct and in an extra array used only for printing.

The global variables of the kalman filter are now static variables inside it's
function and the output is also stored in the datapoint struct.

The datapoint struct is also used for printing to reduce memory usage
and make everything more consistent.

The main thing disabled here is just the mpu6050 and with it the state
machine.

(This commit has also been used to record some test data)